+++ src/parts/component.js
@@ -62,20 +62,30 @@ export function makeComponent(node, element) {
         ctx.writeLine(`let $class = {}`);
     }));
 
+    head.push(xNode('namespace', {
+        $cond: () => passOption.namespace
+    }, ctx => {
+        ctx.writeLine('let namespace = {};');
+    }));
+
     let _boundEvents = {};
     const boundEvent = (name) => {
         if(!_boundEvents[name]) _boundEvents[name] = forwardAllEvents ? 1 : 0;
         _boundEvents[name]++;
     }
 
-
     if(node.body && node.body.length) {
         let slots = {};
+        let namespaces = [];
         let defaultSlot = {
             name: 'default',
             type: 'slot'
         }
         defaultSlot.body = node.body.filter(n => {
+            if(n.type == 'node' && n.name[0] == '^') {
+                namespaces.push(n);
+                return false;
+            }
             if(n.type != 'slot') return true;
             let rx = n.value.match(/^\#slot:(\S+)/);
             if(rx) n.name = rx[1];
@@ -148,6 +158,29 @@ export function makeComponent(node, element) {
                         ctx.build(data.setters);
                     });
                     ctx.writeLine(`};`);
+                });
+                ctx.writeLine(`}`);
+            }));
+        });
+
+        namespaces.forEach(n => {
+            passOption.namespace = true;
+            let block = this.buildBlock({body: [n]}, {inline: true, oneElement: 'el', bindAttributes: true});
+            let name = n.name.slice(1);
+            head.push(xNode('namespace', {
+                name: name || 'default',
+                source: block.source
+            }, (ctx, data) => {
+                ctx.writeLine(`namespace.${data.name} = function(el) {`);
+                ctx.goIndent(() => {
+                    ctx.writeLine(`let $childCD = $cd.new();`);
+                    ctx.writeLine(`{`);
+                    ctx.goIndent(() => {
+                        ctx.writeLine(`let $cd = $childCD;`);
+                        ctx.build(data.source);
+                    });
+                    ctx.writeLine(`}`);
+                    ctx.writeLine(`return () => {$childCD.destroy();}`);
                 });
                 ctx.writeLine(`}`);
             }));
@@ -392,6 +425,7 @@ export function makeComponent(node, element) {
     if(forwardAllEvents || passOption.events) options.push('events');
     if(passOption.slots) options.push('slots');
     if(passOption.class) options.push('$class');
+    if(passOption.namespace) options.push('namespace');
 
     let result = xNode('component', {
         el: element.bindName(),