+++ src/parts/prop.js
@@ -323,6 +323,14 @@ export function bindProp(prop, node, element) {
             });
             return {bind};
         }
+    } else if(name[0] == '^') {
+        this.require('apply');
+        return {bind: xNode('bindNamespace', {
+            name: name.slice(1) || 'default',
+            el: element.bindName()
+        }, (ctx, n) => {
+            ctx.writeLine(`$runtime.attachNamespace($component, $cd, '${n.name}', ${n.el});`)
+        })};
     } else {
         if(prop.value && prop.value.indexOf('{') >= 0) {
             this.require('apply');