+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * local config "malina.config.js"
 * plugin sass/scss
 * key "auto" for each-block
+* namespaces for components
 
 ## 0.5.x